Bind API tokens into a separate database
